nuru-system/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Date/Time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# HTTP requests
httpx==0.25.2

# Logging
structlog==23.2.0

# CORS
python-cors==1.7.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1